# openApi.yaml
...
swagger: "2.0"
info:
  title: Users and documents API
  description: This API is used in document-validation APP
  version: 1.1.0
host: document-validator.appspot.com
...
basePath: /v1
schemes:
  - https
paths:
...
  /users/{id}:
    get:
      summary: Return a user based on ID.
      operationId: getUserById
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        200:
          description: OK
        500:
          description: Error getting user
      security:
        - api_key: []
    put:
      summary: Edit a user
      operationId: editUser
      parameters:
        - in: body
          name: user
          schema:
            $ref: "#/definitions/User"
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: User edited successfully
        500:
          description: Error editing user
      security:
        - api_key: []
    delete:
      summary: delete a user
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: User deleted successfully
        500:
          description: Error deleting user
      security:
        - api_key: []
  /users/username/{username}:
    get:
      summary: Return a user based on username.
      operationId: getUserByUsername
      parameters:
        - in: path
          name: username
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: OK
        500:
          description: Error getting user
      security:
        - api_key: []
...
  /documents/{id}:
    delete:
      summary: delete a document
      operationId: deleteDocument
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Document deleted successfully
        500:
          description: Error deleting document
      security:
        - api_key: []
    get:
      summary: Return the list of documents linked to the userId passed.
      operationId: getDocumentsByUserId
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        200:
          description: OK
        500:
          description: Error getting user
      security:
        - api_key: []
...
definitions:
...
  User:
    type: "object"
    properties:
      username:
        type: "string"
      fullName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phones:
        type: "array"
        items:
          type: "string"
...
securityDefinitions:
  # This section configures basic authentication with an API key.
  api_key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
# [END securityDef]